Complete refactoring, removal of duplicate code, new interface for signals.
